requirements.txt
Django==1.8.7
django-cors-headers==1.1.0
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.3.1
ecdsa==0.13
Fabric==1.10.2
freeze==1.0.9
paramiko==1.16.0
pycrypto==2.6.1
PyYAML==3.11
requests==2.8.1
six==1.10.0
wheel==0.24.0
